% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_test.R
\name{chisq.test1}
\alias{chisq.test1}
\title{chi-square test statistic}
\usage{
chisq.test1(x, y)
}
\arguments{
\item{x}{Sample to be tested}

\item{y}{Sample to be tested}
}
\value{
chi-square test statistic,p-value and degree of freedom.
}
\description{
give two sample vectors,generate list of chi-square test statistic,degree of freedom and p-value.
}
\examples{
\dontrun{
a=31:35;
b=c(31,33,35,37,39);
m<-cbind(a,b)
chisq.test1(a,b)
}
}
