% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvm_test.R
\name{cvm}
\alias{cvm}
\title{Cram′er-von Mises test}
\usage{
cvm(x, y, N)
}
\arguments{
\item{x}{the sample to be tested}

\item{y}{the sample to be tested}

\item{N}{the number of permutation}
}
\value{
test p-value
}
\description{
generating Cram′er-von Mises test p-value,which the null hypothesis that the two samples come from the same distribution.
}
