% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rayleigh.R
\name{MC.Phi}
\alias{MC.Phi}
\title{random numbers which are Rayleighed distribution}
\usage{
MC.Phi(x, sigma, R = 10000, antithetic = TRUE)
}
\arguments{
\item{x}{the initiation sequence}

\item{sigma}{the parameter of Rayleigh density}

\item{R}{the number of replication samples}

\item{antithetic}{the method which generates the cdf.default is TRUE}
}
\value{
a random vector that the number is length(x)
}
\description{
the function for generating distribution function random numbers
}
\examples{
\dontrun{
x <- seq(.1,2.5,length=5);
set.seed(123)
MC1<- MC.Phi(x,sigma=2,anti=FALSE) #for (X1+X2)/2 which X1,X2 is independent
MC2<- MC.Phi(x,sigma=2,anti=TRUE)  #for antithetic variables (X+X')/2
print(round(rbind(x, MC1, MC2, Phi),5))
}
}
